requirements.txt
attrs==20.3.0
Automat==20.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
distro==1.5.0
docker==4.4.4
docker-compose==1.28.6
dockerpty==0.4.1
docopt==0.6.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
lxml==4.6.3
MarkupSafe==1.1.1
numpy==1.20.2
pandas==1.2.4
paramiko==2.7.2
parsel==1.6.0
Protego==0.1.16
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
queuelib==1.5.0
requests==2.25.1
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.6
texttable==1.6.3
Twisted==21.2.0
urllib3==1.26.4
w3lib==1.22.0
websocket-client==0.58.0
Werkzeug==1.0.1
zope.interface==5.3.0

streamlit~=0.80.0
matplotlib~=3.4.1
altair~=4.1.0